supporting over 2G file
git-svn-id: https://svn.qdecoder.org/qhttpd/trunk@45 915dc90f-ddda-dc11-960a-0015c5f490bc

+++ src/http_header.c
@@ -65,7 +65,7 @@ bool httpHeaderHasStr(Q_ENTRY *entries, const char *pszName, const char *pszValu
 	return false;
 }
 
-bool httpHeaderParseRange(const char *pszRangeHeader, size_t nFilesize, off_t *pnRangeOffset1, off_t *pnRangeOffset2, size_t *pnRangeSize) {
+bool httpHeaderParseRange(const char *pszRangeHeader, off_t nFilesize, off_t *pnRangeOffset1, off_t *pnRangeOffset2, off_t *pnRangeSize) {
 	if(pszRangeHeader == NULL) return false;
 
 	char *pszRange = strdup(pszRangeHeader);
